app/backend/requirements.txt
flask
Keras
numpy
scikit-image
scikit-learn[alldeps]
tensorflow
Pillow
